Remove IrMathPrinter
It's identical to IrPrinter.
