+++ csrc/ir/iostream.h
@@ -37,40 +37,6 @@ inline std::ostream& indent(std::ostream& os, int indent_size) {
   return os;
 }
 
-//! Define pretty printing functions for IR nodes
-//!
-//! This class is intended for debug printing, so it attempts
-//! to handle invalid states as well.
-//!
-class IrPrinter {
- public:
-  explicit IrPrinter(std::ostream& os, int indent_size = 0)
-      : os_(os), indent_size_(indent_size) {}
-  virtual ~IrPrinter() = default;
-
-  void resetIndent() {
-    indent_size_ = 0;
-  }
-
-  bool printInline() const {
-    return print_inline_;
-  }
-
-  virtual void handle(const Fusion* f);
-  virtual void handle(const kir::Kernel* kernel);
-  virtual void handle(const hir::HostIrContainer* host_ir_container);
-
- protected:
-  std::ostream& os() {
-    return os_;
-  }
-
- private:
-  std::ostream& os_;
-  bool print_inline_ = false;
-  int indent_size_ = 0;
-};
-
 NVF_API std::ostream& operator<<(std::ostream& os, const Statement& stmt);
 NVF_API std::ostream& operator<<(std::ostream& os, const Statement* stmt);
