+++ csrc/ir/printer.h
@@ -16,8 +16,8 @@ namespace nvfuser {
 
 //! Prints computation Fusion IR nodes
 //!
-//! IrMathPrinter and IrTransformPrinter allow the splitting up of fusion print
-//! functions. IrMathPrinter as its name implies focuses solely on what tensor
+//! IrPrinter and IrTransformPrinter allow the splitting up of fusion print
+//! functions. IrPrinter as its name implies focuses solely on what tensor
 //! computations are taking place. Resulting TensorView math will reflect the
 //! series of split/merge/computeAts that have taken place, however these
 //! nodes will not be displayed in what is printed. IrTransformPrinter does not
@@ -27,22 +27,38 @@ namespace nvfuser {
 //! of a fusion.
 //
 //! \sa IrTransformPrinter
-//!
-class IrMathPrinter : public IrPrinter {
+class IrPrinter {
  public:
-  IrMathPrinter(std::ostream& os) : IrPrinter(os) {}
+  explicit IrPrinter(std::ostream& os, int indent_size = 0)
+      : os_(os), indent_size_(indent_size) {}
+  virtual ~IrPrinter() = default;
 
-  using IrPrinter::handle;
-
-  void handle(const Fusion* f) override {
-    IrPrinter::handle(f);
+  void resetIndent() {
+    indent_size_ = 0;
   }
+
+  bool printInline() const {
+    return print_inline_;
+  }
+
+  virtual void handle(const Fusion* f);
+  virtual void handle(const kir::Kernel* kernel);
+  virtual void handle(const hir::HostIrContainer* host_ir_container);
+
+ protected:
+  std::ostream& os() {
+    return os_;
+  }
+
+ private:
+  std::ostream& os_;
+  bool print_inline_ = false;
+  int indent_size_ = 0;
 };
 
 //! Prints transformation (schedule) Fusion IR nodes
 //!
-//! \sa IrMathPrinter
-//!
+//! \sa IrPrinter
 class IrTransformPrinter : public IrPrinter {
  public:
   IrTransformPrinter(std::ostream& os) : IrPrinter(os) {}
